Move test into separate class, to allow for better separation between mock and real thing
